Initial commit of spatial_mus
spatial_mus is a django application that was built to workout the kinks
in developing an api endpoint (or points) that will dynamically return
the appropriate management unit for samples.  Samples need to be
analyzed at different spatial scales depending on the analysis - local
to basin wide.  When samples are saved, their coordinates are used to
associate them with all of the known management units.  The type of
management unit that is returned is specified as a query parameter (e.g.
- 'qma', 'aa', 'ltrz' ect.)