Basic working views, urls and serializers
This commit includes views, urls and serializers for each of our four
models - management unit, management unit type, fn011 and fn121. The
fn121 view has a custom queryset that prefetches the appropriate
management unit depending on the value of an optional url query
parameter.  The serializers has a custom method field that returns the
slug of the management unit or none if one does not exist.